requirements.txt
streamlit
PyPDF2
langchain
langchain-community
faiss-cpu
sentence-transformers
groq
phidata
fastapi
python-dotenv

requests==2.31.0
selenium
beautifulsoup4
webdriver-manager


nltk